NEW: directory and files restructuring
