+++ src/components/UIComponents/PlioList.vue
@@ -11,7 +11,7 @@
 
 <script>
 import { mapActions, mapState } from 'vuex';
-import PlioThumbnail from "@/components/PlioThumbnail.vue";
+import PlioThumbnail from "@/components/UIComponents/PlioThumbnail.vue";
 
 export default {
   name: "PlioList",
