+++ src/components/UIComponents/PlioThumbnail.vue
