+++ src/components/UIComponents/tutorial/ProceedButtonPointer.vue
