+++ src/components/UIComponents/tutorial/StartButtonPointer.vue
